# api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BookmarkHive API",
        "description": "Decentralized social bookmarking service based on the Activity Pub protocol.",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://bookmarkhive.test",
            "description": "API Dev Server"
        },
        {
            "url": "https://api2.bookmarkhive.test",
            "description": "API 2 Dev Server"
        }
    ],
    "paths": {
...
        "/users/me/bookmarks/{id}": {
            "get": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Get own bookmark by ID",
                "description": "Returns a specific bookmark owned by the authenticated user.",
                "operationId": "getOwnBookmark",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bookmark ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookmark details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookmarkShowPrivate"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "404": {
                        "description": "Bookmark not found or not owned by user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Delete a bookmark",
                "description": "Permanently deletes a bookmark (and all its versions/history) owned by the authenticated user.",
                "operationId": "deleteBookmark",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bookmark ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Bookmark deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "404": {
                        "description": "Bookmark not found or not owned by user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Update a bookmark",
                "description": "Updates an existing bookmark. URL cannot be changed after creation (URL field in request body is ignored). Tags are preserved if not included in the request body.",
                "operationId": "updateBookmark",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bookmark ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Bookmark update data",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookmarkUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bookmark updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookmarkShowPrivate"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "404": {
                        "description": "Bookmark not found or not owned by user"
                    },
                    "422": {
                        "description": "Validation error - invalid data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "invalid_data": {
                                        "summary": "Validation error",
                                        "value": {
                                            "error": {
                                                "code": 422,
                                                "message": "Unprocessable Content"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/me/bookmarks/{id}/history": {
            "get": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Get bookmark history",
                "description": "Returns the history of outdated bookmarks with the same normalized URL as the given bookmark. The current bookmark is not included in the history. Results are ordered by creation date (newest first).",
                "operationId": "getBookmarkHistory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bookmark ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of outdated bookmarks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "collection": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookmarkShowPrivate"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "404": {
                        "description": "Bookmark not found or not owned by user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/profile/{username}/bookmarks": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "List public bookmarks of a user",
                "description": "Returns a paginated collection of public bookmarks owned by the specified user. Supports filtering by tags (only public tags) and cursor-based pagination. Default page size is 24 items, ordered by creation date (newest first).",
                "operationId": "listPublicBookmarks",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "Username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Comma-separated list of tag slugs to filter by (only public tags)",
                        "schema": {
                            "type": "string",
                            "example": "tag-one,tag-two"
                        }
                    },
                    {
                        "name": "after",
                        "in": "query",
                        "description": "Cursor for pagination - bookmark ID to fetch results after",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of public bookmarks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "collection": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookmarkShowPublic"
                                            }
                                        },
                                        "prevPage": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "nextPage": {
                                            "description": "URL for next page if available",
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "total": {
                                            "type": "integer",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BookmarkShowPrivate": {
                "description": "Bookmark object with owner-level details",
                "properties": {
                    "id": {
                        "description": "Bookmark ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "createdAt": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "title": {
                        "description": "Bookmark title",
                        "type": "string"
                    },
                    "url": {
                        "description": "Bookmark URL",
                        "type": "string",
                        "format": "uri"
                    },
                    "domain": {
                        "description": "Extracted domain from URL",
                        "type": "string"
                    },
                    "account": {
                        "$ref": "#/components/schemas/AccountShowPublic"
                    },
                    "isPublic": {
                        "description": "Whether the bookmark is public",
                        "type": "boolean"
                    },
                    "tags": {
                        "description": "Associated tags",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "mainImage": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/FileObject"
                            }
                        ],
                        "nullable": true,
                        "description": "Main image file object",
                        "type": "object"
                    },
                    "archive": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/FileObject"
                            }
                        ],
                        "nullable": true,
                        "description": "Archive file object",
                        "type": "object"
                    },
                    "instance": {
                        "description": "Instance host where the bookmark was created",
                        "type": "string"
                    },
                    "@iri": {
                        "description": "IRI of the bookmark resource",
                        "type": "string",
                        "format": "iri"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token obtained from POST /auth endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}